Input validation for cli flags
