+++ Makefile
@@ -84,11 +84,11 @@ dry-run-control-plane: kustomize manifests
 
 .PHONY: build
 build: generate fmt vet ## Build manager binary.
-	go build -o bin/manager main.go
+	go build -o bin/manager main.go flags.go
 
 .PHONY: run
 run: manifests generate fmt vet ## Run a controller from your host.
-	go run ./main.go
+	go run ./main.go ./flags.go
 
 .PHONY: docker-build
 docker-build: ## Build docker image with the manager.
